+++ app/Http/Controllers/ArticleController.php
@@ -15,8 +15,12 @@ class ArticleController extends Controller
      */
     public function index()
     {
-        $data['articles'] = Article::with('evenement')->get();
-        return view("article.index", $data);
+        if (auth()->user()->role === 'admin') {
+            $data['articles'] = Article::with('evenement')->get();
+            return view("article.index", $data);
+        } else {
+            return redirect()->back();
+        }
     }
 
     /**
@@ -26,8 +30,12 @@ class ArticleController extends Controller
      */
     public function create()
     {
-        $data['events'] = Evenement::all();
-        return view("article.create", $data);
+        if (auth()->user()->role === 'admin') {
+            $data['events'] = Evenement::all();
+            return view("article.create", $data);
+        } else {
+            return redirect()->back();
+        }
     }
 
     /**
@@ -38,28 +46,32 @@ class ArticleController extends Controller
      */
     public function store(Request $request)
     {
-        // $request->validate([
-        //     'titre' => 'required|alpha',
-        //     'contenu' => 'required',
-        //     'description' => 'required'
-        // ], [
-        //     'title.required' => 'Title is required',
-        //     'contenu.required' => 'Contenu is required',
-        //     'description.required' => 'Description is required',
-        // ]);
-        $imageName = time() . '.' . $request->image->extension();
-        $request->image->move(public_path('articleImages'), $imageName);
+        if (auth()->user()->role === 'admin') {
+            // $request->validate([
+            //     'titre' => 'required|alpha',
+            //     'contenu' => 'required',
+            //     'description' => 'required'
+            // ], [
+            //     'title.required' => 'Title is required',
+            //     'contenu.required' => 'Contenu is required',
+            //     'description.required' => 'Description is required',
+            // ]);
+            $imageName = time() . '.' . $request->image->extension();
+            $request->image->move(public_path('articleImages'), $imageName);
 
-        $article = new Article;
-        $article->titre = $request->title;
-        $article->image = $request->image;
-        $article->description = $request->description;
-        $article->contenu = $request->content;
-        $article->evenement_id = $request->event_id;
-        $article->save();
+            $article = new Article;
+            $article->titre = $request->title;
+            $article->image = $request->image;
+            $article->description = $request->description;
+            $article->contenu = $request->content;
+            $article->evenement_id = $request->event_id;
+            $article->save();
 
-        return redirect()->route('article.index')
-            ->with('success', 'Article has been created successfully.');
+            return redirect()->route('articles.index')
+                ->with('success', 'Article has been created successfully.');
+        } else {
+            return redirect()->back();
+        }
     }
 
     /**
@@ -70,8 +82,12 @@ class ArticleController extends Controller
      */
     public function show($id)
     {
-        $article = Article::find($id);
-        return view('article.show',compact('article'));
+        if (auth()->user()->role === 'admin') {
+            $article = Article::find($id);
+            return view('article.show', compact('article'));
+        } else {
+            return redirect()->back();
+        }
     }
 
     /**
@@ -82,9 +98,13 @@ class ArticleController extends Controller
      */
     public function edit($id)
     {
-        $events = Evenement::all();
-        $article = Article::find($id);
-        return view('article.edit', compact('article', 'events'));
+        if (auth()->user()->role === 'admin') {
+            $events = Evenement::all();
+            $article = Article::find($id);
+            return view('article.edit', compact('article', 'events'));
+        } else {
+            return redirect()->back();
+        }
     }
 
     /**
@@ -96,20 +116,24 @@ class ArticleController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $article = Article::find($id);
-        if (!is_string($request->image)) {
-            $imageName = time() . '.' . $request->image->extension();
-            $request->image->move(public_path('articleImages'), $imageName);
-            $article->image = $imageName;
-        }
-        $article->titre = $request->title;
-        $article->description = $request->description;
-        $article->contenu = $request->content;
-        $article->evenement_id = $request->event_id;
-        $article->save();
+        if (auth()->user()->role === 'admin') {
+            $article = Article::find($id);
+            if (!is_string($request->image)) {
+                $imageName = time() . '.' . $request->image->extension();
+                $request->image->move(public_path('articleImages'), $imageName);
+                $article->image = $imageName;
+            }
+            $article->titre = $request->title;
+            $article->description = $request->description;
+            $article->contenu = $request->content;
+            $article->evenement_id = $request->event_id;
+            $article->save();
 
-        return redirect()->route('articles.index')
-            ->with('success', 'Article Has Been updated successfully');
+            return redirect()->route('articles.index')
+                ->with('success', 'Article Has Been updated successfully');
+        } else {
+            return redirect()->back();
+        }
     }
 
     /**
@@ -120,9 +144,13 @@ class ArticleController extends Controller
      */
     public function destroy($id)
     {
-        $article = Article::where('id', $id)->firstOrFail();
-        $article->delete();
-        return redirect()->route('articles.index')
-            ->with('success', 'Article has been deleted successfully');
+        if (auth()->user()->role === 'admin') {
+            $article = Article::where('id', $id)->firstOrFail();
+            $article->delete();
+            return redirect()->route('articles.index')
+                ->with('success', 'Article has been deleted successfully');
+        } else {
+            return redirect()->back();
+        }
     }
 }
